+++ IPBanCore/Core/IPBan/IPBanLogFileManager.cs
@@ -109,8 +109,8 @@ namespace DigitalRuby.IPBanCore
                             MaxFileSizeBytes = newFile.MaxFileSize,
                             PathAndMask = pathAndMask,
                             PingIntervalMilliseconds = (service.ManualCycle ? 0 : newFile.PingInterval),
-                            RegexFailure = IPBanConfig.ParseRegex(newFile.FailedLoginRegex, true),
-                            RegexSuccess = IPBanConfig.ParseRegex(newFile.SuccessfulLoginRegex, true),
+                            RegexFailure = IPBanRegexParser.ParseRegex(newFile.FailedLoginRegex, true),
+                            RegexSuccess = IPBanRegexParser.ParseRegex(newFile.SuccessfulLoginRegex, true),
                             RegexFailureTimestampFormat = newFile.FailedLoginRegexTimestampFormat,
                             RegexSuccessTimestampFormat = newFile.SuccessfulLoginRegexTimestampFormat,
                             Source = newFile.Source,
