+++ IPBanCore/Core/IPBan/IPBanService.cs
@@ -145,166 +145,6 @@ namespace DigitalRuby.IPBanCore
         }
 
         /// <summary>
-        /// Get an ip address and user name out of text using regex. Regex may contain groups named source_[sourcename] to override the source.
-        /// </summary>
-        /// <param name="regex">Regex</param>
-        /// <param name="text">Text</param>
-        /// <param name="timestampFormat">Timestamp format</param>
-        /// <param name="eventType">Event type</param>
-        /// <param name="info">Info</param>
-        /// <param name="dns">Dns lookup to resolve ip addresses</param>
-        /// <returns>Set of matches from text</returns>
-        public static IEnumerable<IPAddressLogEvent> GetIPAddressEventsFromRegex(Regex regex, string text,
-            string timestampFormat = null, IPAddressEventType eventType = IPAddressEventType.FailedLogin,
-            string info = null, IDnsLookup dns = null)
-        {
-            const string customSourcePrefix = "source_";
-
-            // if no regex or no text, we are done
-            if (regex is null || string.IsNullOrWhiteSpace(text))
-            {
-                yield break;
-            }
-
-            // remove control chars
-            text = new string(text.Where(c => c == '\n' || c == '\t' || !char.IsControl(c)).ToArray());
-
-            // go through all the matches and pull out event info
-            var matches = regex.Matches(text);
-            foreach (var match in matches.Cast<Match>())
-            {
-                string userName = null;
-                string ipAddress = null;
-                string foundSource = null;
-                string logData = null;
-                DateTime timestamp = default;
-
-                // check for a user name
-                if (string.IsNullOrWhiteSpace(userName))
-                {
-                    var userNameGroup = match.Groups["username"];
-                    if (userNameGroup != null && userNameGroup.Success)
-                    {
-                        userName ??= userNameGroup.Value.Trim(regexTrimChars);
-                    }
-                    else
-                    {
-                        // sometimes user names are base64, like smtp logs
-                        userNameGroup = match.Groups["username_base64"];
-                        if (userNameGroup != null && userNameGroup.Success)
-                        {
-                            // attempt to decode base64 and get the actual user name
-                            var base64UserName = userNameGroup.Value;
-                            Span<byte> bytes = stackalloc byte[256];
-                            if (Convert.TryFromBase64String(base64UserName, bytes, out int bytesWritten))
-                            {
-                                var base64DecodedUserName = System.Text.Encoding.UTF8.GetString(bytes[..bytesWritten]);
-                                userName ??= base64DecodedUserName.Trim(regexTrimChars);
-                            }
-                        }
-                    }
-                }
-
-                // check for source
-                if (string.IsNullOrWhiteSpace(foundSource))
-                {
-                    var sourceGroup = match.Groups["source"];
-                    if (sourceGroup != null && sourceGroup.Success)
-                    {
-                        foundSource = sourceGroup.Value.Trim(regexTrimChars);
-                    }
-                }
-
-                // check for groups with a custom source name
-                foreach (var group in match.Groups.Cast<Group>())
-                {
-                    if (group.Success &&
-                        group.Name != null &&
-                        string.IsNullOrWhiteSpace(foundSource) &&
-                        group.Name.StartsWith(customSourcePrefix))
-                    {
-                        foundSource = group.Name[customSourcePrefix.Length..];
-                    }
-                }
-
-                // check for timestamp group
-                var timestampGroup = match.Groups["timestamp"];
-                if (timestampGroup != null && timestampGroup.Success)
-                {
-                    string toParse = timestampGroup.Value.Trim(regexTrimChars);
-                    if (string.IsNullOrWhiteSpace(timestampFormat) ||
-                        !DateTime.TryParseExact(toParse, timestampFormat.Trim(), CultureInfo.InvariantCulture,
-                            DateTimeStyles.AssumeLocal | DateTimeStyles.AdjustToUniversal, out timestamp))
-                    {
-                        DateTime.TryParse(toParse, CultureInfo.InvariantCulture,
-                            DateTimeStyles.AssumeLocal | DateTimeStyles.AdjustToUniversal, out timestamp);
-                    }
-                }
-
-                // check if the regex had an ipadddress group
-                var ipAddressGroup = match.Groups["ipaddress"];
-                if (ipAddressGroup is null || !ipAddressGroup.Success)
-                {
-                    ipAddressGroup = match.Groups["ipaddress_exact"];
-                }
-                if (ipAddressGroup != null && ipAddressGroup.Success && !string.IsNullOrWhiteSpace(ipAddressGroup.Value))
-                {
-                    string tempIPAddress = ipAddressGroup.Value.Trim();
-
-                    // in case of IP:PORT format, try a second time, stripping off the :PORT, saves having to do this in all
-                    //  the different ip regex.
-                    int lastColon = tempIPAddress.LastIndexOf(':');
-                    bool isValidIPAddress = IPAddress.TryParse(tempIPAddress, out IPAddress tmp);
-                    if (isValidIPAddress || (lastColon >= 0 && IPAddress.TryParse(tempIPAddress[..lastColon], out tmp)))
-                    {
-                        ipAddress = tmp.ToString();
-                    }
-
-                    // if we are parsing anything as ip address (including dns names)
-                    if (string.IsNullOrWhiteSpace(ipAddress) &&
-                        dns != null &&
-                        ipAddressGroup.Name == "ipaddress" &&
-                        tempIPAddress != Environment.MachineName &&
-                        tempIPAddress != "-")
-                    {
-                        // Check Host by name
-                        Logger.Info("Parsing as IP failed for {0}, info: {1}. Checking dns...", tempIPAddress, info);
-                        try
-                        {
-                            IPAddress[] ipAddresses = dns.GetHostAddressesAsync(tempIPAddress).Sync();
-                            if (ipAddresses != null && ipAddresses.Length > 0)
-                            {
-                                ipAddress = ipAddresses.FirstOrDefault().ToString();
-                                Logger.Info("Dns result '{0}' = '{1}'", tempIPAddress, ipAddress);
-                            }
-                        }
-                        catch
-                        {
-                            Logger.Info("Parsing as dns failed '{0}'", tempIPAddress);
-                        }
-                    }
-                }
-
-                // check for log data
-                if (string.IsNullOrWhiteSpace(logData))
-                {
-                    var logDataGroup = match.Groups["log"];
-                    if (logDataGroup is not null && logDataGroup.Success)
-                    {
-                        logData = logDataGroup.Value;
-                    }
-                }
-
-                // see if there is a repeat indicator in the message
-                int repeatCount = ExtractRepeatCount(match, text);
-
-                // return an event for this match
-                yield return new IPAddressLogEvent(ipAddress, userName, foundSource, repeatCount,
-                    eventType, timestamp, logData: logData);
-            }
-        }
-
-        /// <summary>
         /// Write a new config file
         /// </summary>
         /// <param name="xml">Xml of the new config file</param>
