+++ IPBanCore/ipban.config
@@ -927,6 +927,12 @@ Login failed for user 'cenptuser'. Reason: Failed to open the explicitly specifi
     <add key="BlacklistRegex" value=""/>
 
     <!--
+	When a parser finds a user name, you can truncate the username to any of the characters in this value. Set to empty to not truncate.
+	Very useful when parsing email addresses and hackers are trying same user name at multiple different domain names.
+	-->
+    <add key="TruncateUserNameChars" value="@"/>
+	  
+    <!--
       Comma separated list of user names that are allowed. UserNameWhitelistMinimumEditDistance is checked for user names not in the list
       to determine whether a failed user name should be banned. Case is ignored.
       Empty entries are allowed if you want to allow empty user names.
