+++ IPBanTests/IPBanLogFileIntegrationTests.cs
@@ -88,7 +88,7 @@ namespace DigitalRuby.IPBanTests
             {
                 Assert.AreEqual("MSExchange", successfulEvents[i].Source);
                 Assert.AreEqual(IPAddressEventType.SuccessfulLogin, successfulEvents[i].Type);
-                Assert.AreEqual("user@example.com", successfulEvents[i].UserName);
+                Assert.AreEqual("user", successfulEvents[i].UserName);
             }
 
             // 37.49.225.153, UserName: p.kurowicki@gios.gov.pl, Source: MSExchange, Count: 1, Type: FailedLogin, Timestamp: 6/26/2021 3:01:36 PM}
@@ -103,7 +103,7 @@ namespace DigitalRuby.IPBanTests
                 Assert.AreEqual("MSExchange", failedEvents[i].Source);
                 if (i != failedEvents.Count - 2)
                 {
-                    Assert.AreEqual("user@example.com", failedEvents[i].UserName);
+                    Assert.AreEqual("user", failedEvents[i].UserName);
                 }
                 else
                 {
