+++ IPBanTests/IPBanLogFileParserTests.cs
@@ -295,9 +295,9 @@ namespace DigitalRuby.IPBanTests
                 EventHandler = this,
                 Source = source,
                 PathAndMask = pathAndMask,
-                RegexFailure = IPBanConfig.ParseRegex(failureRegex, true),
+                RegexFailure = IPBanRegexParser.ParseRegex(failureRegex, true),
                 RegexFailureTimestampFormat = failureRegexTimestampFormat,
-                RegexSuccess = IPBanConfig.ParseRegex(successRegex, true),
+                RegexSuccess = IPBanRegexParser.ParseRegex(successRegex, true),
                 RegexSuccessTimestampFormat = successRegexTimestampFormat
             };
             LogFileScanner scanner = new IPBanLogFileScanner(options);
